Refactor admin app routers (#10)
* Delete spec file that was unused

* Add fixture files used by tests

* Add data migrations with test data

* Add sample questions

* Bump version to v0.2.0

* Bump version to v0.2.0

* Update docs

* Update .gitignore

* Update user model

* Update .gitignore

* number of questions are now 10

* changing the expectation

* changing the expectation for warm-up

* Add newlines at end of file

* fixed tests

* Change let to const

* Split out route logic to controllers and relocate helper functions

* Standard JS compliance

* Same principle for route index

* Refactor route helper to service

* Fix comment

* Remove next call from service and include JSDoc changes
