+++ admin/app.js
@@ -31,7 +31,7 @@ autoIncrement.initialize(mongoose.connection)
 
 const index = require('./routes/index')
 const testDeveloper = require('./routes/test-developer')
-const school = require('./routes/school')
+const admin = require('./routes/admin')
 
 const app = express()
 const helpers = require('./helpers')(app)
@@ -144,7 +144,7 @@ app.use(function (req, res, next) {
 
 app.use('/', index)
 app.use('/test-developer', testDeveloper)
-app.use('/school', school)
+app.use('/school', admin)
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
